Image guide to locate "Use this Template" button
I found fellow ambassador asking question where to locate the "Use this Template" button on the repository. 

I had similar question for the first 20 minutes going through the repo and was able to figure out that I didn't signed in.

So, I added few instructions and a visual guide (image of where the button is located).